requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
websockets>=11.0.3
pydantic>=2.4.2
face-recognition>=1.3.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
pytest>=7.4.2
pytest-asyncio>=0.21.1
python-dotenv>=1.0.0